Many changes, not complete, but shelving this for now
Added/consolidated files, added monsters class and then a library consisting of a few instances of that class for different mob types. Learning how to reference those instances outside of the file.

Have more, but don't want to get stuck on trying to perfect his right now. Might revisit it down the road.